+++ Changelog.md
@@ -1,12 +1,18 @@
 #West Wind Toolkit Change Log
 
-### Version 2.57
+### Version 2.58
 *not released yet*
 
 * **New HttpUtils.JsonRequest**<br/>
 Added a new HttpUtils class with a JsonRequest() and JsonRequestAsync() 
 methods to handle calling JSON services and automatically sending
 and receiving of JSON data.
+
+* **New HttpUtils class**<br />
+Added static HttpUtils class to make it easy to make Http Requests
+and specifically to make JSON service calls that can automatically
+serialize and deserialize data. Class also includes a simple HTTP
+retrieval routine.
 
 * **AppConfiguration support for Nested Property Encryption**<br />
 You can now specify nested fields for encryption in the provider's
@@ -27,6 +33,12 @@ styles and attributes. This update adds support for monitoring
 attributes and uses the new MutationObserver DOM Api that works with
 all modern browsers. Slight interface change passing parameters
 as an `options` parameter
+
+* **Split .NET 4.0 and 4.5 targets for Westwind.Utilities**<br />
+Create seperate net40 target for .NET 4.0 compatible output of
+Westwind.Utilities while moving forward to 4.5 with most of the
+code. Start integrating a number of async features into new and
+existing utility classes.
 
 ### Version 2.56
 *October 2nd, 2014*